admin/requirements.txt
-i https://pypi.org/simple
appnope==0.1.0 ; sys_platform == 'darwin'
backcall==0.1.0
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
colorlog==3.1.4
decorator==4.3.0
django-appconf==1.0.2
django-extensions==2.1.3
django-grappelli==2.12.1
django-imagekit==4.0.2
django-storages[google]==1.7.1
django-taggit==0.23.0
django-tinymce==2.7.0
django==2.1.3
google-api-core==1.5.2
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-storage==1.13.0 ; extra == 'google'
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
gunicorn==19.9.0
idna==2.7
ipdb==0.11
ipython-genutils==0.2.0
ipython==7.1.1
jedi==0.13.1
mysqlclient==1.3.13
netifaces==0.10.7
parso==0.3.1
pexpect==4.6.0 ; sys_platform != 'win32'
pickleshare==0.7.5
pilkit==2.0
pillow==5.3.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
pyasn1-modules==0.2.2
pyasn1==0.4.4
pygments==2.2.0
pytz==2018.7
requests==2.20.1
rsa==4.0
six==1.11.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
